refactor Rename KeeperContract to ContractWrapperBase
Unified with squid-js
