+++ squid_py/keeper/market.py
@@ -3,17 +3,17 @@ import logging
 from squid_py.constants import (
     OCEAN_MARKET_CONTRACT
 )
-from squid_py.keeper.keeper_contract import (
-    KeeperContract,
+from squid_py.keeper.contract_base import (
+    ContractWrapperBase,
 )
 
 DEFAULT_GAS_LIMIT = 400000
 
 
-class Market(KeeperContract):
+class Market(ContractWrapperBase):
 
     def __init__(self, web3_helper, contract_path, address):
-        KeeperContract.__init__(self, web3_helper, OCEAN_MARKET_CONTRACT, 'market', contract_path, address)
+        ContractWrapperBase.__init__(self, web3_helper, OCEAN_MARKET_CONTRACT, 'market', contract_path, address)
         self._defaultGas = DEFAULT_GAS_LIMIT
 
     # call functions (costs no gas)
