+++ squid_py/utils/web3_helper.py
@@ -18,26 +18,26 @@ class Web3Helper(object):
     def __init__(self, web3):
         self._web3 = web3
 
-    def load(self, contract_file, name, contract_path, contract_address):
-        """Retrieve a tuple with the concise contract and the contract definition."""
-        contract_filename = os.path.join(contract_path, "{}.json".format(contract_file))
-        try:
-            valid_address = self._web3.toChecksumAddress(contract_address)
-        except ValueError as e:
-            raise OceanInvalidContractAddress("Invalid contract address for keeper contract '{}'".format(name))
-        except Exception as e:
-            raise e
-
-        with open(contract_filename, 'r') as abi_definition:
-            abi = json.load(abi_definition)
-            concise_cont = self._web3.eth.contract(
-                address=valid_address,
-                abi=abi['abi'],
-                ContractFactoryClass=ConciseContract)
-            contract = self._web3.eth.contract(
-                address=valid_address,
-                abi=abi['abi'])
-            return concise_cont, contract, contract_address
+    # def load(self, contract_file, name, contract_path, contract_address):
+    #     """Retrieve a tuple with the concise contract and the contract definition."""
+    #     contract_filename = os.path.join(contract_path, "{}.json".format(contract_file))
+    #     try:
+    #         valid_address = self._web3.toChecksumAddress(contract_address)
+    #     except ValueError as e:
+    #         raise OceanInvalidContractAddress("Invalid contract address for keeper contract '{}'".format(name))
+    #     except Exception as e:
+    #         raise e
+    #
+    #     with open(contract_filename, 'r') as abi_definition:
+    #         abi = json.load(abi_definition)
+    #         concise_cont = self._web3.eth.contract(
+    #             address=valid_address,
+    #             abi=abi['abi'],
+    #             ContractFactoryClass=ConciseContract)
+    #         contract = self._web3.eth.contract(
+    #             address=valid_address,
+    #             abi=abi['abi'])
+    #         return concise_cont, contract, contract_address
 
     def sign(self, account_address, message):
         return self._web3.eth.sign(account_address, message)
