+++ tests/test_ocean.py
@@ -33,8 +33,8 @@ def get_keeper_path(path=''):
 def test_ocean_contracts():
     os.environ['CONFIG_FILE'] = 'config_local.ini'
     os.environ['KEEPER_URL'] = 'http://0.0.0.0:8545'
-    ocean = Ocean()
-    assert ocean.contracts.token is not None
+    ocean = Ocean(os.environ['CONFIG_FILE'])
+    assert ocean.keeper.token is not None
     assert ocean.keeper_url == os.environ['KEEPER_URL']
